Move ActivityLifecycleCallbacksAdapter to tests
